+++ src/core/helpers.py
@@ -24,10 +24,10 @@
 
 import os
 from sys import exit
-from core import commands, modules
+from core import builtin_functions, modules
 from core import parser
 
-class Helpers(commands.Commands):
+class Helpers(builtin_functions.BuiltinFunctions):
     """ Partial of program object functions """
 
     def raise_variable_error(self, varname: str, op: dict):
@@ -84,7 +84,7 @@ class Helpers(commands.Commands):
             do_raise_error = False
             try:
                 if self.all_vars()[self.current_namespace() + varname] != None:
-                    op = self.states[-1]['operations'][self.states[-1]['current_step']]
+                    op = self.states[-1]['commands'][self.states[-1]['current_step']]
                     do_raise_error = True
             except:
                 pass
@@ -123,14 +123,14 @@ class Helpers(commands.Commands):
             while i > 0:
                 self.states.pop()
                 i -= 1
-            self.states[-1]['current_step'] = len(self.states[-1]['operations']) * 2
+            self.states[-1]['current_step'] = len(self.states[-1]['commands']) * 2
             self.exit_code = exit_code
 
     def pashmak_eval(self, code):
         """ Runs the pashmak code from string """
         # run the code
-        code_operations = parser.parse(code, filepath='<eval>')
-        self.exec_func(code_operations, False)
+        code_commands = parser.parse(code, filepath='<eval>')
+        self.exec_func(code_commands, False)
 
     def current_namespace(self):
         """ Returns current namespace """
@@ -141,4 +141,4 @@ class Helpers(commands.Commands):
 
     def signal_handler(self, signal_code, frame):
         """ Raise error when signal exception raised """
-        self.raise_error('Signal', str(signal_code), self.states[-1]['operations'][self.states[-1]['current_step']])
+        self.raise_error('Signal', str(signal_code), self.states[-1]['commands'][self.states[-1]['current_step']])