+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Pashmak Changelog
 
-## 0.6, 0.6-beta2
+## 0.6, 0.6-beta3
 
 - created **Class** system
 - change default `RuntimeError` name to the real python exception name in exception system
@@ -20,6 +20,7 @@
 - fixed lot of interpreter internal bugs and some optimizations
 - added inline function calling feature, for example `println %{ my_func_or_command %{ another_command_as_arg }% }%`
 - removed `$var = ^ command` syntax(use `$var = %{ command }%` instead of that)
+- added general variable `$__ismain__`
 
 ## 0.5.4 (2020-12-6)